Added license header file so that every new file we add automatically gets the header.
This is provided by the "License Header Manager" extension in the VS Gallery for those that want to install it.

Note that nobody needs to have the VSIX in order for the solution to work, compile, etc. It just makes it far easier to manage the headers at design-time (there's even a project-wide "Add header" thing to add it to any file that doesn't have it already). A must-have for me :)