Refactored ConferenceController,
in order to remove a couple of leaky abstractions. Taking a dependency on
Func<IViewRepository> indicates knowledge about a particular
implementation.

In addition, the implementation violated the Liskov Substitution Principle
by attempting to downcast the created repository to IDisposable. This was
also a leaky abstraction.

Much to my surprise, NO UNIT TESTS BROKE because of this refactoring.
Either this project has surprisingly loosely coupled tests, or else the
coverage could be better.